move nethermind to a single file
