+++ docker/test_env/nethermind.go
@@ -12,6 +12,30 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
 )
 
+func TestEth1WithNethermind(t *testing.T) {
+	l := logging.GetTestLogger(t)
+
+	builder := NewEthereumNetworkBuilder()
+	cfg, err := builder.
+		WithConsensusType(ConsensusType_PoW).
+		WithExecutionLayer(ExecutionLayer_Nethermind).
+		Build()
+	require.NoError(t, err, "Builder validation failed")
+
+	net, _, err := cfg.Start()
+	require.NoError(t, err, "Couldn't start PoW network")
+
+	c, err := blockchain.ConnectEVMClient(net, l)
+	require.NoError(t, err, "Couldn't connect to the evm client")
+
+	address := common.HexToAddress("0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1")
+	err = sendAndCompareBalances(testcontext.Get(t), c, address)
+	require.NoError(t, err, fmt.Sprintf("balance wasn't correctly updated when %s network", net.Name))
+
+	err = c.Close()
+	require.NoError(t, err, "Couldn't close the client")
+}
+
 func TestEth2WithPrysmAndNethermind(t *testing.T) {
 	l := logging.GetTestLogger(t)
