+++ docker/test_env/nethermind_pos.go
@@ -2,6 +2,7 @@ package test_env
 
 import (
 	"context"
+	"encoding/hex"
 	"fmt"
 	"os"
 	"strings"
@@ -20,11 +21,13 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/docker"
 	"github.com/smartcontractkit/chainlink-testing-framework/logging"
 	"github.com/smartcontractkit/chainlink-testing-framework/mirror"
+	"github.com/smartcontractkit/chainlink-testing-framework/utils/templates"
 )
 
 const defaultNethermindPosImage = "nethermind/nethermind:1.25.1"
+const defaultNethermindPoaImage = "nethermind/nethermind:1.16.0"
 
-type NethermindPos struct {
+type Nethermind struct {
 	EnvComponent
 	ExternalHttpUrl      string
 	InternalHttpUrl      string
@@ -33,14 +36,15 @@ type NethermindPos struct {
 	InternalExecutionURL string
 	ExternalExecutionURL string
 	generatedDataHostDir string
+	chainConfg           *EthereumChainConfig
 	consensusLayer       ConsensusLayer
 	l                    zerolog.Logger
 	t                    *testing.T
 }
 
-func NewNethermindPos(networks []string, generatedDataHostDir string, consensusLayer ConsensusLayer, opts ...EnvComponentOption) (*NethermindPos, error) {
+func NewNethermindPos(networks []string, generatedDataHostDir string, consensusLayer ConsensusLayer, opts ...EnvComponentOption) (*Nethermind, error) {
 	parts := strings.Split(defaultNethermindPosImage, ":")
-	g := &NethermindPos{
+	g := &Nethermind{
 		EnvComponent: EnvComponent{
 			ContainerName:    fmt.Sprintf("%s-%s", "nethermind-pos", uuid.NewString()[0:8]),
 			Networks:         networks,
@@ -59,14 +63,40 @@ func NewNethermindPos(networks []string, generatedDataHostDir string, consensusL
 	return g, nil
 }
 
-func (g *NethermindPos) WithTestInstance(t *testing.T) ExecutionClient {
+func NewNethermindPoa(networks []string, chainConfg *EthereumChainConfig, opts ...EnvComponentOption) (*Nethermind, error) {
+	parts := strings.Split(defaultNethermindPoaImage, ":")
+	g := &Nethermind{
+		EnvComponent: EnvComponent{
+			ContainerName:    fmt.Sprintf("%s-%s", "nethermind-poa", uuid.NewString()[0:8]),
+			Networks:         networks,
+			ContainerImage:   parts[0],
+			ContainerVersion: parts[1],
+		},
+		chainConfg: chainConfg,
+		l:          logging.GetTestLogger(nil),
+	}
+	for _, opt := range opts {
+		opt(&g.EnvComponent)
+	}
+	// if the internal docker repo is set then add it to the version
+	g.EnvComponent.ContainerImage = mirror.AddMirrorToImageIfSet(g.EnvComponent.ContainerImage)
+	return g, nil
+}
+
+func (g *Nethermind) WithTestInstance(t *testing.T) ExecutionClient {
 	g.l = logging.GetTestLogger(t)
 	g.t = t
 	return g
 }
 
-func (g *NethermindPos) StartContainer() (blockchain.EVMNetwork, error) {
-	r, err := g.getContainerRequest()
+func (g *Nethermind) StartContainer() (blockchain.EVMNetwork, error) {
+	var r *tc.ContainerRequest
+	var err error
+	if g.consensusLayer != "" {
+		r, err = g.getPosContainerRequest()
+	} else {
+		r, err = g.getPowContainerRequest()
+	}
 	if err != nil {
 		return blockchain.EVMNetwork{}, err
 	}
@@ -97,9 +127,14 @@ func (g *NethermindPos) StartContainer() (blockchain.EVMNetwork, error) {
 	if err != nil {
 		return blockchain.EVMNetwork{}, err
 	}
-	executionPort, err := ct.MappedPort(context.Background(), NatPort(ETH2_EXECUTION_PORT))
-	if err != nil {
-		return blockchain.EVMNetwork{}, err
+
+	if g.consensusLayer != "" {
+		executionPort, err := ct.MappedPort(context.Background(), NatPort(ETH2_EXECUTION_PORT))
+		if err != nil {
+			return blockchain.EVMNetwork{}, err
+		}
+		g.InternalExecutionURL = FormatHttpUrl(g.ContainerName, ETH2_EXECUTION_PORT)
+		g.ExternalExecutionURL = FormatHttpUrl(host, executionPort.Port())
 	}
 
 	g.Container = ct
@@ -107,11 +142,14 @@ func (g *NethermindPos) StartContainer() (blockchain.EVMNetwork, error) {
 	g.InternalHttpUrl = FormatHttpUrl(g.ContainerName, DEFAULT_EVM_NODE_HTTP_PORT)
 	g.ExternalWsUrl = FormatWsUrl(host, wsPort.Port())
 	g.InternalWsUrl = FormatWsUrl(g.ContainerName, DEFAULT_EVM_NODE_WS_PORT)
-	g.InternalExecutionURL = FormatHttpUrl(g.ContainerName, ETH2_EXECUTION_PORT)
-	g.ExternalExecutionURL = FormatHttpUrl(host, executionPort.Port())
 
 	networkConfig := blockchain.SimulatedEVMNetwork
-	networkConfig.Name = fmt.Sprintf("Simulated Ethereum-PoS (nethermind + %s)", g.consensusLayer)
+	if g.consensusLayer != "" {
+		networkConfig.Name = fmt.Sprintf("Simulated Ethereum-PoS (nethermind + %s)", g.consensusLayer)
+	} else {
+		networkConfig.Name = "Simulated Ethereum-PoA (nethermind)"
+		networkConfig.GasEstimationBuffer = 100_000_000_000
+	}
 	networkConfig.URLs = []string{g.ExternalWsUrl}
 	networkConfig.HTTPURLs = []string{g.ExternalHttpUrl}
 
@@ -121,39 +159,162 @@ func (g *NethermindPos) StartContainer() (blockchain.EVMNetwork, error) {
 	return networkConfig, nil
 }
 
-func (g *NethermindPos) GetInternalExecutionURL() string {
+func (g *Nethermind) getPowContainerRequest() (*tc.ContainerRequest, error) {
+	keystoreDir, err := os.MkdirTemp("", "keystore")
+	if err != nil {
+		return nil, err
+	}
+
+	generatedData, err := generateKeystoreAndExtraData(keystoreDir)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisJsonStr, err := templates.NethermindPoAGenesisJsonTemplate{
+		ChainId:     fmt.Sprintf("%d", g.chainConfg.ChainID),
+		AccountAddr: RootFundingAddr,
+		ExtraData:   fmt.Sprintf("0x%s", hex.EncodeToString(generatedData.extraData)),
+	}.String()
+	if err != nil {
+		return nil, err
+	}
+	genesisFile, err := os.CreateTemp("", "genesis_json")
+	if err != nil {
+		return nil, err
+	}
+	_, err = genesisFile.WriteString(genesisJsonStr)
+	if err != nil {
+		return nil, err
+	}
+
+	// create empty cfg file since if we don't pass any
+	// default mainnet.cfg will be used
+	noneCfg, err := os.CreateTemp("", "none.cfg")
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = noneCfg.WriteString("{}")
+	if err != nil {
+		return nil, err
+	}
+
+	passFile, err := os.CreateTemp("", "password.txt")
+	if err != nil {
+		return nil, err
+	}
+
+	err = os.WriteFile(passFile.Name(), []byte(""), 0600)
+	if err != nil {
+		return nil, err
+	}
+
+	rootFile, err := os.CreateTemp(keystoreDir, RootFundingAddr)
+	if err != nil {
+		return nil, err
+	}
+	_, err = rootFile.WriteString(RootFundingWallet)
+	if err != nil {
+		return nil, err
+	}
+
+	return &tc.ContainerRequest{
+		Name:            g.ContainerName,
+		Image:           g.GetImageWithVersion(),
+		Networks:        g.Networks,
+		AlwaysPullImage: true,
+		// ImagePlatform: "linux/x86_64",  //don't even try this on Apple Silicon, the node won't start due to .NET error
+		ExposedPorts: []string{NatPortFormat(DEFAULT_EVM_NODE_HTTP_PORT), NatPortFormat(DEFAULT_EVM_NODE_WS_PORT)},
+		WaitingFor: tcwait.ForAll(
+			tcwait.ForLog("Nethermind initialization completed").
+				WithStartupTimeout(120 * time.Second).
+				WithPollInterval(1 * time.Second),
+		),
+		Cmd: []string{
+			"--config=/none.cfg",
+			"--Init.ChainSpecPath=/chainspec.json",
+			"--Init.DiscoveryEnabled=false",
+			"--Init.WebSocketsEnabled=true",
+			fmt.Sprintf("--JsonRpc.WebSocketsPort=%s", DEFAULT_EVM_NODE_WS_PORT),
+			"--JsonRpc.Enabled=true",
+			"--JsonRpc.EnabledModules=net,consensus,eth,subscribe,web3,admin,trace,txpool",
+			"--JsonRpc.Host=0.0.0.0",
+			fmt.Sprintf("--JsonRpc.Port=%s", DEFAULT_EVM_NODE_HTTP_PORT),
+			"--KeyStore.KeyStoreDirectory=/keystore",
+			fmt.Sprintf("--KeyStore.BlockAuthorAccount=%s", generatedData.minerAccount.Address.Hex()),
+			fmt.Sprintf("--KeyStore.UnlockAccounts=%s", generatedData.minerAccount.Address.Hex()),
+			"--KeyStore.PasswordFiles=/password.txt",
+			"--Mining.Enabled=true",
+			"--Init.PeerManagerEnabled=false",
+			"--HealthChecks.Enabled=true", // default slug /health
+		},
+		Files: []tc.ContainerFile{
+			{
+				HostFilePath:      genesisFile.Name(),
+				ContainerFilePath: "/chainspec.json",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      noneCfg.Name(),
+				ContainerFilePath: "/none.cfg",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      passFile.Name(),
+				ContainerFilePath: "/password.txt",
+				FileMode:          0644,
+			},
+		},
+		HostConfigModifier: func(hostConfig *container.HostConfig) {
+			hostConfig.Mounts = append(hostConfig.Mounts, mount.Mount{
+				Type:     mount.TypeBind,
+				Source:   keystoreDir,
+				Target:   "/keystore",
+				ReadOnly: false,
+			})
+		},
+		LifecycleHooks: []tc.ContainerLifecycleHooks{
+			{
+				PostStarts: g.PostStartsHooks,
+				PostStops:  g.PostStopsHooks,
+			},
+		},
+	}, nil
+}
+
+func (g *Nethermind) GetInternalExecutionURL() string {
 	return g.InternalExecutionURL
 }
 
-func (g *NethermindPos) GetExternalExecutionURL() string {
+func (g *Nethermind) GetExternalExecutionURL() string {
 	return g.ExternalExecutionURL
 }
 
-func (g *NethermindPos) GetInternalHttpUrl() string {
+func (g *Nethermind) GetInternalHttpUrl() string {
 	return g.InternalHttpUrl
 }
 
-func (g *NethermindPos) GetInternalWsUrl() string {
+func (g *Nethermind) GetInternalWsUrl() string {
 	return g.InternalWsUrl
 }
 
-func (g *NethermindPos) GetExternalHttpUrl() string {
+func (g *Nethermind) GetExternalHttpUrl() string {
 	return g.ExternalHttpUrl
 }
 
-func (g *NethermindPos) GetExternalWsUrl() string {
+func (g *Nethermind) GetExternalWsUrl() string {
 	return g.ExternalWsUrl
 }
 
-func (g *NethermindPos) GetContainerName() string {
+func (g *Nethermind) GetContainerName() string {
 	return g.ContainerName
 }
 
-func (g *NethermindPos) GetContainer() *tc.Container {
+func (g *Nethermind) GetContainer() *tc.Container {
 	return &g.Container
 }
 
-func (g *NethermindPos) getContainerRequest() (*tc.ContainerRequest, error) {
+func (g *Nethermind) getPosContainerRequest() (*tc.ContainerRequest, error) {
 	// create empty cfg file since if we don't pass any
 	// default mainnet.cfg will be used
 	noneCfg, err := os.CreateTemp("", "none.cfg")
@@ -224,11 +385,14 @@ func (g *NethermindPos) getContainerRequest() (*tc.ContainerRequest, error) {
 	}, nil
 }
 
-func (g *NethermindPos) WaitUntilChainIsReady(ctx context.Context, waitTime time.Duration) error {
+func (g *Nethermind) WaitUntilChainIsReady(ctx context.Context, waitTime time.Duration) error {
+	if g.consensusLayer == "" {
+		return nil
+	}
 	waitForFirstBlock := tcwait.NewLogStrategy("Improved post-merge block").WithPollInterval(1 * time.Second).WithStartupTimeout(waitTime)
 	return waitForFirstBlock.WaitUntilReady(ctx, *g.GetContainer())
 }
 
-func (g *NethermindPos) GetContainerType() ContainerType {
+func (g *Nethermind) GetContainerType() ContainerType {
 	return ContainerType_Nethermind
 }
